Trying to register pid, still untested

+++ kernel/src/arch/x86_64/syscall.rs
@@ -499,7 +499,7 @@ fn handle_fileio(
             #[cfg(feature = "exokernel")]
             {
                 let mut client = kcb.arch.rpc_client.lock();
-                return match client.as_mut().unwrap().fio_close(pid as u64, fd) {
+                return match client.as_mut().unwrap().fio_close(pid, fd) {
                     Ok(a) => Ok(a),
                     Err(err) => Err(err.into()),
                 };
